src/requirements.txt
click==8.1.3
Flask==2.2.2
PyYaml==6.0
flask-cors==3.0.10
Flask-BasicAuth==0.2.0
python-dateutil==2.8.2
tornado==6.2
marko==1.2.1